test: add tests to DeleteUser service

+++ src/users/services/delete-user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { UserRepository } from '../repositories/user.repository';
 import { User } from '../entities/user.entity';
 
@@ -11,7 +11,11 @@ export class DeleteUserService {
   constructor(private readonly repository: UserRepository) {}
 
   public async execute({ id }: IRequest): Promise<User> {
-    const user = await this.repository.findOneOrFail(id);
+    const user = await this.repository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
 
     const oldUser = { ...user };
     await this.repository.remove(user);
